+++ Editor/AssemblyInfo.cs
@@ -1,3 +1,3 @@
 ﻿using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Unity.2D.SpriteShape.Tests.EditorTests")]
-[assembly: InternalsVisibleTo("InternalsVisible.ToDynamicProxyGenAssembly2")]
+[assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")] // Castle DynamicProxy, used by NSubstitute to proxy the internal IAngleRange* interfaces